requirements.txt
pandas
PySocks
python-dateutil
python-dotenv
python-engineio
python-socketio
pathtools
itsdangerous
Jinja2
MarkupSafe
Pillow
SQLAlchemy
virtualenv
Werkzeug
psycopg2
pyowm
gunicorn
alembic
bleach
blinker
chardet
Click
PyJWT
psycopg2
python-dotenv
newsapi-python
Flask
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SocketIO
Flask-SQLAlchemy
Flask-SeaSurf
Flask-Caching
Flask-Cors
Flask-HTTPAuth
googletrans
validator-collection
gevent
gevent-websocket
nltk
beautifulsoup4
dnspython
ChatterBot
chatterbot-corpus
marshmallow